backend/requirements.txt
Flask
python-dotenv
kiteconnect
pymysql
pyjwt
